♻️ Refactor: Chatting component 분리
